@@ -13,6 +13,10 @@ using System.Reflection;
 using System.Runtime.CompilerServices;
 using Newtonsoft.Json.Schema;
 using NbCore.Platform.Graphics;
+using NbCore.IO;
+using OpenTK.Windowing.Common.Input;
+using Microsoft.CodeAnalysis;
+using System.Reflection.Metadata.Ecma335;
 
 namespace NbCore
 {
@@ -311,6 +315,12 @@ namespace NbCore
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public void RegisterEntity(NbMaterial mat)
         {
+            if (mat is null)
+            {
+                Log("Null Material", LogVerbosityLevel.WARNING);
+                return;
+            }
+                
             if (GetSystem<EntityRegistrySystem>().RegisterEntity(mat))
             {
                 GetSystem<RenderingSystem>().MaterialMgr.AddMaterial(mat);
@@ -604,11 +614,13 @@ namespace NbCore
             //Load Resources 
             LoadDefaultResources();
             AddDefaultShaderSources();
-            AddDefaultShaderConfigs();
-
-            CompileMainShaders();
-            AddDefaultMaterials();
-            AddDefaultPrimitives();
+            //AddDefaultShaderConfigs();
+            
+            LoadAssets();
+            
+            //CompileMainShaders();
+            //AddDefaultMaterials();
+            //AddDefaultPrimitives();
             
             Log("Initialized", LogVerbosityLevel.INFO);
         }
@@ -644,7 +656,7 @@ namespace NbCore
                 }
             }
 
-            DirectoryInfo dirInfo = new("Shaders");
+            DirectoryInfo dirInfo = new("Assets/Shaders/Source");
             if (!dirInfo.Exists)
                 return;
             
@@ -662,574 +674,68 @@ namespace NbCore
 
         }
 
-
-        private void AddDefaultShaderConfigs()
+        private void LoadAssets()
         {
-            //Create Debug Shader Config
-            //Debug Shader
-            NbShaderConfig conf;
-            conf = CreateShaderConfig(GetShaderSourceByFilePath("Shaders/Simple_VSEmpty.glsl"),
-                                      GetShaderSourceByFilePath("Shaders/Simple_FSEmpty.glsl"),
-                                      GetShaderSourceByFilePath("Shaders/Simple_GS.glsl"), null, null,
-                                      NbShaderMode.DEFFERED, "Debug", true);
-            RegisterEntity(conf);
+            Dictionary<Type, List<string>> fileDict = new();
+            fileDict[typeof(NbMesh)] = new();
+            fileDict[typeof(NbMeshData)] = new();
+            fileDict[typeof(NbMaterial)] = new();
+            fileDict[typeof(NbShaderConfig)] = new();
+            fileDict[typeof(NbShader)] = new();
 
+            //Fetch files
+            Directory.CreateDirectory("Assets");
+            var files = Directory.EnumerateFiles("Assets", "*.nb*", SearchOption.AllDirectories);
 
-            //Test Shader
-            conf = CreateShaderConfig(GetShaderSourceByFilePath("Shaders/Simple_VSEmpty.glsl"),
-                                      GetShaderSourceByFilePath("Shaders/Test_fs.glsl"),
-                                      null, null, null,
-                                      NbShaderMode.DEFFERED, "Test", true);
-            RegisterEntity(conf);
-
-            //UberShader Shader
-            conf = CreateShaderConfig(GetShaderSourceByFilePath("Shaders/Simple_VS.glsl"),
-                                      GetShaderSourceByFilePath("Shaders/ubershader_fs.glsl"),
-                                      null, null, null,
-                                      NbShaderMode.DEFFERED, "UberShader_Deferred", true);
-            RegisterEntity(conf);
-
-            //UberShader Lit Shader
-            conf = CreateShaderConfig(GetShaderSourceByFilePath("Shaders/Simple_VS.glsl"),
-                                      GetShaderSourceByFilePath("Shaders/ubershader_fs.glsl"),
-                                      null, null, null,
-                                      NbShaderMode.DEFFERED | NbShaderMode.LIT, "UberShader_Deferred_Lit", true);
-            RegisterEntity(conf);
-
-            //UNLIT
-            conf = CreateShaderConfig(GetShaderSourceByFilePath("Shaders/light_pass_VS.glsl"),
-                                      GetShaderSourceByFilePath("Shaders/light_pass_FS.glsl"),
-                                      null, null, null,
-                                      NbShaderMode.DEFFERED, "LightPass_Unlit_Forward", true); ;
-            RegisterEntity(conf);
-
-            //UNLIT
-            conf = CreateShaderConfig(GetShaderSourceByFilePath("Shaders/light_pass_VS.glsl"),
-                                      GetShaderSourceByFilePath("Shaders/light_pass_FS.glsl"),
-                                      null, null, null,
-                                      NbShaderMode.FORWARD, "LightPass_Unlit_Forward", true); ;
-            RegisterEntity(conf);
-
-
-            //LIT
-            conf = CreateShaderConfig(GetShaderSourceByFilePath("Shaders/light_pass_VS.glsl"),
-                                      GetShaderSourceByFilePath("Shaders/light_pass_FS.glsl"),
-                                      null, null, null,
-                                      NbShaderMode.FORWARD | NbShaderMode.LIT, "LightPass_Lit_Forward", true); ;
-            RegisterEntity(conf);
-
-
-            //GAUSSIAN HORIZONTAL BLUR SHADER
-            conf = CreateShaderConfig(GetShaderSourceByFilePath("Shaders/Gbuffer_VS.glsl"),
-                                      GetShaderSourceByFilePath("Shaders/gaussian_horizontalBlur_FS.glsl"),
-                                      null, null, null,
-                                      NbShaderMode.FORWARD, "Horizontal_Gaussian_Blur", true);
-            RegisterEntity(conf);
-
-
-            //GAUSSIAN VERTICAL BLUR SHADER
-            conf = CreateShaderConfig(GetShaderSourceByFilePath("Shaders/Gbuffer_VS.glsl"),
-                                      GetShaderSourceByFilePath("Shaders/gaussian_verticalBlur_FS.glsl"),
-                                      null, null, null,
-                                      NbShaderMode.DEFAULT, "Vertical_Gaussian_Blur", true);
-            RegisterEntity(conf);
-
-            //Brightness Extraction Shader
-            conf = CreateShaderConfig(GetShaderSourceByFilePath("Shaders/Gbuffer_VS.glsl"),
-                                      GetShaderSourceByFilePath("Shaders/brightness_extract_shader_fs.glsl"),
-                                      null, null, null,
-                                      NbShaderMode.FORWARD, "Brightness_Extract", true);
-            RegisterEntity(conf);
-
-            //ADDITIVE BLEND
-            conf = CreateShaderConfig(GetShaderSourceByFilePath("Shaders/Gbuffer_VS.glsl"),
-                                      GetShaderSourceByFilePath("Shaders/additive_blend_fs.glsl"),
-                                      null, null, null,
-                                      NbShaderMode.FORWARD, "Additive_Blend", true);
-            RegisterEntity(conf);
-
-            //FXAA
-            conf = CreateShaderConfig(GetShaderSourceByFilePath("Shaders/fxaa_shader_vs.glsl"),
-                                      GetShaderSourceByFilePath("Shaders/fxaa_shader_fs.glsl"),
-                                      null, null, null,
-                                      NbShaderMode.DEFAULT, "FXAA", true);
-            RegisterEntity(conf);
-
-            //TONE MAPPING + GAMMA CORRECTION
-            conf = CreateShaderConfig(GetShaderSourceByFilePath("Shaders/Gbuffer_VS.glsl"),
-                                      GetShaderSourceByFilePath("Shaders/tone_mapping_fs.glsl"),
-                                      null, null, null,
-                                      NbShaderMode.FORWARD, "ToneMapping", true);
-            RegisterEntity(conf);
-
-            //INV TONE MAPPING
-            conf = CreateShaderConfig(GetShaderSourceByFilePath("Shaders/Gbuffer_VS.glsl"),
-                                      GetShaderSourceByFilePath("Shaders/inv_tone_mapping_fs.glsl"),
-                                      null, null, null,
-                                      NbShaderMode.FORWARD, "InverseToneMapping", true);
-            RegisterEntity(conf);
-
-            //BWOIT SHADER
-            conf = CreateShaderConfig(GetShaderSourceByFilePath("Shaders/Gbuffer_VS.glsl"),
-                                      GetShaderSourceByFilePath("Shaders/bwoit_shader_fs.glsl"),
-                                      null, null, null,
-                                      NbShaderMode.FORWARD, "BWOIT", true);
-            RegisterEntity(conf);
-
-            //Text Shaders
-            conf = CreateShaderConfig(GetShaderSourceByFilePath("Shaders/Text_VS.glsl"),
-                                      GetShaderSourceByFilePath("Shaders/Text_FS.glsl"),
-                                      null, null, null,
-                                      NbShaderMode.FORWARD, "Text", true);
-            RegisterEntity(conf);
-
-            //Pass Shader
-            conf = CreateShaderConfig(GetShaderSourceByFilePath("Shaders/Gbuffer_VS.glsl"),
-                                      GetShaderSourceByFilePath("Shaders/PassThrough_FS.glsl"),
-                                      null, null, null,
-                                      NbShaderMode.FORWARD, "Passthrough", true);
-            RegisterEntity(conf);
-
-            //Red Shader
-            conf = CreateShaderConfig(GetShaderSourceByFilePath("Shaders/Gbuffer_VS.glsl"),
-                                      GetShaderSourceByFilePath("Shaders/RedFill.glsl"),
-                                      null, null, null,
-                                      NbShaderMode.FORWARD, "RedFill");
-            RegisterEntity(conf);
-
-        }
-
-
-        private void CompileMainShaders()
-        {
-            //Populate shader list
-
-            NbShaderConfig shader_conf;
-            NbShader shader;
-
-            //LIT
-            shader = new()
+            foreach (string file in files)
             {
-                Type = NbShaderType.LIGHT_PASS_LIT_SHADER,
-                IsGeneric = true
-            };
-
-            shader.SetShaderConfig(GetShaderConfigByName("LightPass_Lit_Forward"));
-            Callbacks.Assert(CompileShader(shader), "Error on shader compilation");
-            RegisterEntity(shader);
-
-            //GAUSSIAN HORIZONTAL BLUR SHADER
-            shader = new()
-            {
-                IsGeneric = true
-            };
-            shader.SetShaderConfig(GetShaderConfigByName("Horizontal_Gaussian_Blur"));
-            CompileShader(shader);
-            RegisterEntity(shader);
-
-            //GAUSSIAN VERTICAL BLUR SHADER
-            shader = new()
-            {
-                IsGeneric = true
-            };
-            shader.SetShaderConfig(GetShaderConfigByName("Vertical_Gaussian_Blur"));
-            CompileShader(shader);
-            RegisterEntity(shader);
-
-            //BRIGHTNESS EXTRACTION SHADER
-            shader = new()
-            {
-                IsGeneric = true
-            };
-
-            shader.SetShaderConfig(GetShaderConfigByName("Brightness_Extract"));
-            CompileShader(shader);
-            RegisterEntity(shader);
-
-            //ADDITIVE BLEND
-            shader = new()
-            {
-                IsGeneric = true
-            };
-            shader.SetShaderConfig(GetShaderConfigByName("Additive_Blend"));
-            CompileShader(shader);
-            RegisterEntity(shader);
-
-            //FXAA
-            shader = new()
-            {
-                IsGeneric = true,
-                Type = NbShaderType.FXAA_SHADER,
-            };
-            shader.SetShaderConfig(GetShaderConfigByName("FXAA"));
-            Callbacks.Assert(CompileShader(shader), "Error on shader compilation");
-            RegisterEntity(shader);
-
-            //TONE MAPPING + GAMMA CORRECTION
-            shader = new()
-            {
-                Type = NbShaderType.TONE_MAPPING,
-                IsGeneric = true
-            };
-            shader.SetShaderConfig(GetShaderConfigByName("ToneMapping"));
-            CompileShader(shader);
-            RegisterEntity(shader);
-
-            //INV TONE MAPPING + GAMMA CORRECTION
-            shader = new()
-            {
-                Type = NbShaderType.INV_TONE_MAPPING,
-                IsGeneric = true
-            };
-            shader.SetShaderConfig(GetShaderConfigByName("InverseToneMapping"));
-            CompileShader(shader);
-            RegisterEntity(shader);
-
-            //BWOIT SHADER
-            shader = new()
-            {
-                IsGeneric = true
-            };
-            shader.SetShaderConfig(GetShaderConfigByName("BWOIT"));
-            CompileShader(shader);
-            RegisterEntity(shader);
-
-            //Text Shaders
-            shader = CreateShader(GetShaderConfigByName("Text"), new List<string> { NbMaterialFlagEnum._NB_DIFFUSE_MAP.ToString() });
-            shader.IsGeneric = true;
-            CompileShader(shader);
-            RegisterEntity(shader);
-
-            //FILTERS - EFFECTS
-
-            //Pass Shader
-            shader = new()
-            {
-                Type = NbShaderType.PASSTHROUGH_SHADER,
-                IsGeneric = true
-            };
-            shader.SetShaderConfig(GetShaderConfigByName("Passthrough"));
-            CompileShader(shader);
-            RegisterEntity(shader);
-
-            //Red Shader
-            shader_conf = GetShaderConfigByName("RedFill");
-            shader = new()
-            {
-                Type = NbShaderType.RED_FILL_SHADER,
-                IsGeneric = true
-            };
-            shader.SetShaderConfig(GetShaderConfigByName("RedFill"));
-            CompileShader(shader);
-            RegisterEntity(shader);
-
-        }
-
-        private void AddDefaultMaterials()
-        {
-            //Cross Material
-            NbMaterial mat;
-            NbShaderConfig config_deferred = GetShaderConfigByName("UberShader_Deferred");
-            NbShaderConfig config_deferred_lit = GetShaderConfigByName("UberShader_Deferred_Lit");
-            NbShader shader;
-
-            mat = new()
-            {
-                Name = "crossMat",
-                IsGeneric = true
-            };
-            mat.AddFlag(NbMaterialFlagEnum._NB_UNLIT);
-            mat.AddFlag(NbMaterialFlagEnum._NB_VERTEX_COLOUR);
-            NbUniform uf = new(NbUniformType.Vector4, "gMaterialColourVec4", 1.0f, 1.0f, 1.0f, 1.0f);
-            uf.ShaderBinding = "mpCustomPerMaterial.uniforms[0]";
-            mat.Uniforms.Add(uf);
-
-            shader = CreateShader(config_deferred, GetMaterialShaderDirectives(mat));
-            Callbacks.Assert(CompileShader(shader), "Error during shader compilation");
-            mat.AttachShader(shader);
-#if DEBUG
-            //Report UBOs
-            GraphicsAPI.ShaderReport(shader);
-#endif
-            RegisterEntity(mat); //Register Material
-
-            //Joint Material
-            mat = new NbMaterial
-            {
-                Name = "jointMat",
-                IsGeneric = true
-            };
-            mat.AddFlag(NbMaterialFlagEnum._NB_UNLIT);
-
-            uf = new(NbUniformType.Vector4, "gMaterialColourVec4", 1.0f, 0.0f, 0.0f, 1.0f);
-            uf.ShaderBinding = "mpCustomPerMaterial.uniforms[0]";
-
-            mat.Uniforms.Add(uf);
-
-            ulong shader_hash = CalculateShaderHash(config_deferred, GetMaterialShaderDirectives(mat));
-            shader = GetShaderByHash(shader_hash);
-            if (shader == null)
-            {
-                shader = CreateShader(config_deferred, GetMaterialShaderDirectives(mat));
-                Callbacks.Assert(CompileShader(shader), "Error during shader compilation");
+                if (file.EndsWith(".nbmesh"))
+                    fileDict[typeof(NbMesh)].Add(file);
+                else if (file.EndsWith(".nbmeshdata"))
+                    fileDict[typeof(NbMeshData)].Add(file);
+                else if (file.EndsWith(".nbmaterial"))
+                    fileDict[typeof(NbMaterial)].Add(file);
+                else if (file.EndsWith(".nbshaderconf"))
+                    fileDict[typeof(NbShaderConfig)].Add(file);
+                else if (file.EndsWith(".nbshader"))
+                    fileDict[typeof(NbShader)].Add(file);
             }
 
-            mat.AttachShader(shader);
-            RegisterEntity(mat);
-
-            //Light Material
-            mat = new()
+            //Load Shader configurations
+            foreach (string file in fileDict[typeof(NbShaderConfig)])
             {
-                Name = "lightMat",
-                IsGeneric = true
-            };
-            mat.AddFlag(NbMaterialFlagEnum._NB_UNLIT);
-
-            uf = new(NbUniformType.Vector4, "gMaterialColourVec4", 1.0f, 1.0f, 0.0f, 1.0f);
-            uf.ShaderBinding = "mpCustomPerMaterial.uniforms[0]";
-
-            mat.Uniforms.Add(uf);
-
-            shader_hash = CalculateShaderHash(config_deferred, GetMaterialShaderDirectives(mat));
-            shader = GetShaderByHash(shader_hash);
-            if (shader == null)
-            {
-                shader = CreateShader(config_deferred, GetMaterialShaderDirectives(mat));
-                Callbacks.Assert(CompileShader(shader), "Error during shader compilation");
+                NbShaderConfig ob = (NbShaderConfig)NbDeserializer.Deserialize(NbDeserializer.DeserializeToToken(file));
+                RegisterEntity(ob);
             }
 
-            mat.AttachShader(shader);
-            RegisterEntity(mat);
-
-            //Default Material
-            mat = new()
+            //Load Shaders
+            foreach (string file in fileDict[typeof(NbShader)])
             {
-                Name = "defaultMat",
-                IsGeneric = true
-            };
-            mat.AddFlag(NbMaterialFlagEnum._NB_UNLIT);
-
-            uf = new(NbUniformType.Vector4, "gMaterialColourVec4", 0.7f, 0.7f, 0.7f, 1.0f);
-            uf.ShaderBinding = "mpCustomPerMaterial.uniforms[0]";
-
-            mat.Uniforms.Add(uf);
-
-            shader_hash = CalculateShaderHash(config_deferred, GetMaterialShaderDirectives(mat));
-            shader = GetShaderByHash(shader_hash);
-            if (shader == null)
-            {
-                shader = CreateShader(config_deferred, GetMaterialShaderDirectives(mat));
-                Callbacks.Assert(CompileShader(shader), "Error during shader compilation");
+                NbShader ob = (NbShader)NbDeserializer.Deserialize(NbDeserializer.DeserializeToToken(file));
+                RegisterEntity(ob);
             }
 
-            mat.AttachShader(shader);
-            RegisterEntity(mat);
-
-
-            //Red  Material
-            mat = new()
+            //Load Materials
+            foreach (string file in fileDict[typeof(NbMaterial)])
             {
-                Name = "redMat"
-            };
-
-            NbUniform uf1 = new(NbUniformType.Vector4, "gMaterialColourVec4", 0.7f, 0.7f, 0.7f, 1.0f);
-            uf1.ShaderBinding = "mpCustomPerMaterial.uniforms[0]";
-            mat.Uniforms.Add(uf1);
-
-
-            shader_hash = CalculateShaderHash(config_deferred, GetMaterialShaderDirectives(mat));
-            shader = GetShaderByHash(shader_hash);
-            if (shader == null)
-            {
-                shader = CreateShader(config_deferred, GetMaterialShaderDirectives(mat));
-                Callbacks.Assert(CompileShader(shader), "Error during shader compilation");
+                NbMaterial ob = (NbMaterial)NbDeserializer.Deserialize(NbDeserializer.DeserializeToToken(file));
+                RegisterEntity(ob);
             }
 
-            mat.AttachShader(shader);
-            RegisterEntity(mat);
-
-
-            //Collision Material
-            mat = new()
+            //Load Mesh Data
+            foreach (string file in fileDict[typeof(NbMeshData)])
             {
-                Name = "collisionMat",
-                IsGeneric = true
-            };
-            mat.AddFlag(NbMaterialFlagEnum._NB_UNLIT);
-
-            uf = new(NbUniformType.Vector4, "gMaterialColourVec4", 0.8f, 0.8f, 0.2f, 1.0f);
-            uf.ShaderBinding = "mpCustomPerMaterial.uniforms[0]";
-
-            mat.Uniforms.Add(uf);
-            shader_hash = CalculateShaderHash(config_deferred, GetMaterialShaderDirectives(mat));
-            shader = GetShaderByHash(shader_hash);
-            if (shader == null)
-            {
-                shader = CreateShader(config_deferred, GetMaterialShaderDirectives(mat));
-                CompileShader(shader);
+                NbMeshData ob = (NbMeshData)NbDeserializer.Deserialize(NbDeserializer.DeserializeToToken(file));
+                GetSystem<RenderingSystem>().MeshDataMgr.Add(ob.Hash, ob);
             }
 
-            mat.AttachShader(shader);
-            RegisterEntity(mat);
-
-        }
-
-        private void AddDefaultPrimitives()
-        {
-            //Setup Primitive Vaos
-
-            //Default quad
-            Quad q = new(1.0f, 1.0f);
-
-            NbMesh mesh = new()
+            //Load Mesh
+            foreach (string file in fileDict[typeof(NbMesh)])
             {
-                Hash = NbHasher.Hash("default_quad"),
-                Data = q.geom.GetMeshData(),
-                MetaData = q.geom.GetMetaData(),
-                Material = GetSystem<RenderingSystem>().MaterialMgr.GetByName("defaultMat")
-            };
-
-            RegisterEntity(mesh);
-            q.Dispose();
-
-            //Default render quad
-            q = new Quad();
-
-            mesh = new()
-            {
-                Hash = NbHasher.Hash("default_renderquad"),
-                Data = q.geom.GetMeshData(),
-                MetaData = q.geom.GetMetaData(),
-            };
-            RegisterEntity(mesh);
-            q.Dispose();
-
-            //Default cross
-            Cross c = new(0.1f, true);
-
-            mesh = new()
-            {
-                Type = NbMeshType.Locator, //Explicitly set as locator mesh
-                Hash = NbHasher.Hash("default_cross"),
-                Data = c.geom.GetMeshData(),
-                MetaData = c.geom.GetMetaData(),
-                Material = GetSystem<RenderingSystem>().MaterialMgr.GetByName("crossMat")
-            };
-
-            RegisterEntity(mesh);
-            c.Dispose();
-
-            //Default cube
-            Box bx = new(1.0f, 1.0f, 1.0f, new NbVector3(1.0f), true);
-
-            mesh = new()
-            {
-                Hash = NbHasher.Hash("default_box"),
-                Data = bx.geom.GetMeshData(),
-                MetaData = bx.geom.GetMetaData(),
-                Material = GetSystem<RenderingSystem>().MaterialMgr.GetByName("defaultMat")
-            };
-            RegisterEntity(mesh);
-            bx.Dispose();
-
-            //Default sphere
-            Sphere sph = new(new NbVector3(0.0f, 0.0f, 0.0f), 100.0f);
-
-            mesh = new()
-            {
-                Hash = NbHasher.Hash("default_sphere"),
-                Data = sph.geom.GetMeshData(),
-                MetaData = sph.geom.GetMetaData(),
-                Material = GetSystem<RenderingSystem>().MaterialMgr.GetByName("defaultMat")
-            };
-
-            RegisterEntity(mesh);
-            sph.Dispose();
-
-            //Light Sphere Mesh
-            Sphere lsph = new(new NbVector3(0.0f, 0.0f, 0.0f), 1.0f);
-
-            mesh = new()
-            {
-                Hash = NbHasher.Hash("default_light_sphere"),
-                Data = lsph.geom.GetMeshData(),
-                MetaData = lsph.geom.GetMetaData(),
-                Material = GetSystem<RenderingSystem>().MaterialMgr.GetByName("lightMat")
-            };
-
-            RegisterEntity(mesh);
-            lsph.Dispose();
-
-            GenerateGizmoParts();
-        }
-
-
-        private void GenerateGizmoParts()
-        {
-            //Translation Gizmo
-            Arrow translation_x_axis = new(0.015f, 0.25f, new NbVector3(1.0f, 0.0f, 0.0f), false, 20);
-            //Move arrowhead up in place
-            NbMatrix4 t = NbMatrix4.CreateRotationZ(MathUtils.radians(90));
-            translation_x_axis.applyTransform(t);
-
-            Arrow translation_y_axis = new(0.015f, 0.25f, new NbVector3(0.0f, 1.0f, 0.0f), false, 20);
-            Arrow translation_z_axis = new(0.015f, 0.25f, new NbVector3(0.0f, 0.0f, 1.0f), false, 20);
-            t = NbMatrix4.CreateRotationX(MathUtils.radians(90));
-            translation_z_axis.applyTransform(t);
-
-            //Generate Geom objects
-            translation_x_axis.geom = translation_x_axis.getGeom();
-            translation_y_axis.geom = translation_y_axis.getGeom();
-            translation_z_axis.geom = translation_z_axis.getGeom();
-
-
-            //GLVao x_axis_vao = translation_x_axis.getVAO();
-            //GLVao y_axis_vao = translation_y_axis.getVAO();
-            //GLVao z_axis_vao = translation_z_axis.getVAO();
-
-
-            //Generate PrimitiveMeshVaos
-            for (int i = 0; i < 3; i++)
-            {
-                string name = "";
-                NbCore.Primitives.Primitive arr = null;
-                switch (i)
-                {
-                    case 0:
-                        arr = translation_x_axis;
-                        name = "default_translation_gizmo_x_axis";
-                        break;
-                    case 1:
-                        arr = translation_y_axis;
-                        name = "default_translation_gizmo_y_axis";
-                        break;
-                    case 2:
-                        arr = translation_z_axis;
-                        name = "default_translation_gizmo_z_axis";
-                        break;
-                }
-
-                NbMesh mesh = new()
-                {
-                    Hash = NbHasher.Hash(name),
-                    Data = arr.geom.GetMeshData(),
-                    MetaData = arr.geom.GetMetaData()
-                };
-
-                RegisterEntity(mesh);
-
-                if (!IsEntityRegistered(mesh))
-                    mesh.Dispose();
-                arr.Dispose();
+                NbMesh ob = (NbMesh)NbDeserializer.Deserialize(NbDeserializer.DeserializeToToken(file));
+                RegisterEntity(ob);
             }
 
-        }
-
-
-        //Main Rendering Routines
-        private void rt_ResizeViewport(int w, int h)
-        {
-            GetSystem<RenderingSystem>().Resize(w, h);
         }
 
 #if DEBUG
@@ -1699,20 +1205,17 @@ namespace NbCore
             shader.Hash = CalculateShaderHash(shader);
             
             shader.IsUpdated?.Invoke(shader);
-            //Register New Shader
-            RegisterEntity(shader);
             return true;
         }
 
         #endregion
-
+        
         #region IO
 
         public void SerializeScene(SceneGraph g, string filepath)
         {
             StreamWriter sw = new(filepath);
             Newtonsoft.Json.JsonTextWriter writer = new Newtonsoft.Json.JsonTextWriter(sw);
-            Newtonsoft.Json.JsonSerializer serializer = new Newtonsoft.Json.JsonSerializer();
             writer.Formatting = Newtonsoft.Json.Formatting.Indented;
 
             writer.WriteStartObject();
@@ -1786,7 +1289,29 @@ namespace NbCore
             writer.WritePropertyName("SHADER_CONFIGS");
             writer.WriteStartArray();
             foreach (NbShaderConfig conf in configs)
-                IO.NbSerializer.Serialize(conf, writer);
+            {
+                if (conf.Path == "")
+                    IO.NbSerializer.Serialize(conf, writer);
+                else
+                {
+                    //Write Config file reference to the main file
+                    writer.WriteStartObject();
+                    writer.WritePropertyName("Config");
+                    writer.WriteValue(conf.Path);
+                    writer.WriteEndObject();
+                    
+                    //Serialize Configuration to new file
+                    Directory.CreateDirectory(conf.Path);
+                    StreamWriter config_sw = new(conf.Path);
+                    Newtonsoft.Json.JsonTextWriter config_writer = new Newtonsoft.Json.JsonTextWriter(config_sw);
+                    config_writer.Formatting = Newtonsoft.Json.Formatting.Indented;
+                    config_writer.WriteStartObject();
+                    IO.NbSerializer.Serialize(conf, config_writer);
+                    config_writer.WriteEndObject();
+                    config_writer.Close();
+                }
+            }
+                
             writer.WriteEndArray();
 
             writer.WritePropertyName("TEXTURES");
@@ -1831,12 +1356,9 @@ namespace NbCore
         {
             //Clear Scene
             EngineRef.ClearActiveSceneGraph();
-
-            StreamReader sr = new(filepath);
-            Newtonsoft.Json.JsonTextReader reader = new Newtonsoft.Json.JsonTextReader(sr);
-            var serializer = new Newtonsoft.Json.JsonSerializer();
-            Newtonsoft.Json.Linq.JObject ob = serializer.Deserialize<Newtonsoft.Json.Linq.JObject>(reader);
-            reader.Close();
+            
+            //Deserialize scene file to a json object
+            Newtonsoft.Json.Linq.JObject ob = NbDeserializer.DeserializeToToken(filepath);
             
             //Parse Shader Sources
             Dictionary<ulong,NbMeshData> meshes = new();
